memperbaiki program kalkulator dan template testing

+++ src/main/java/com/b3/kalkulator/operation/Addition.java
@@ -1,7 +1,7 @@
 package com.b3.kalkulator.operation;
 
 public class Addition {
-    public double calculate(double a, double b) {
+    public double calculate(Integer a, Integer b) {
         return a + b;
     }
 }
